+++ ReleaseNotes.md
@@ -2,9 +2,13 @@
 
 # Release portal-disease
 
+## Release 1.3.5
+- Implementation of §7.3 notification (non nominal)
+
 ## Release 1.3.4
 - Updated Readme license disclaimer
 - Move paste box to core lib
+- Update design of input fields
 
 ## Release 1.3.3
 - changed inputs to outline style